resolve pylint issues
